Vulkan: Move shader/pipeline-related methods to ShaderCache
