+++ Source/Core/VideoBackends/Vulkan/Vulkan.vcxproj
@@ -41,6 +41,7 @@
     <ClCompile Include="FramebufferManager.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="PostProcessing.cpp" />
+    <ClCompile Include="ShaderCache.cpp" />
     <ClCompile Include="TextureConverter.cpp" />
     <ClCompile Include="PerfQuery.cpp" />
     <ClCompile Include="RasterFont.cpp" />
@@ -67,6 +68,7 @@
     <ClInclude Include="FramebufferManager.h" />
     <ClInclude Include="Constants.h" />
     <ClInclude Include="PostProcessing.h" />
+    <ClInclude Include="ShaderCache.h" />
     <ClInclude Include="TextureConverter.h" />
     <ClInclude Include="RasterFont.h" />
     <ClInclude Include="StagingBuffer.h" />
